// apps/backend/src/personal-api-key/core/enums/resource.enum.ts
export enum Resource {
  Logs = 'logs',
  Metrics = 'metrics',
  Monitors = 'monitors',
  Projects = 'projects',
  Clusters = 'clusters',
  ClusterInvites = 'cluster_invites', // team membership management
  NotificationChannels = 'notification_channels',
  PublicDashboards = 'public_dashboards',
  CustomDomains = 'custom_domains',
  IngestKeys = 'ingest_keys',
  AuditLog = 'audit_log',
  Account = 'account',
  Payments = 'payments',
}